+++ plugged/vim-clap/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## [unreleased]
 
+## [0.53] 2024-5-2
+
 - Introduce `remote_sink` in provider in order to implement the sink function on the Rust side, particularly useful for the providers with static list of source like lsp.
 - Add `multi_select` property explicitly in provider, useful for the provider lsp.
 - Add `:ClapAction diagnostics.{first,last,next,prev}` for navigating between all kinds of the diagnostics.
 - Add `:ClapAction diagnostics.{firstHint,lastHint,nextHint,prevHint}` for navigating between the Hint diagnostics.
 - Optimize the grep search performance significantly, 2x performance has been achieved compared to the last release.
+
+### Breaking changes
+
+- Rename plugin `cursorword` to `word-highlighter` and added the new feature of highlighting keywords like TODO.
 
 ## [0.52] 2024-2-29